backend/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
PyJWT
passlib[bcrypt]
python-multipart
httpx
python-dotenv
gradio_client